refactor: minor tweaks and optimizations
