+++ examples/mod_cyclonedx-go-v0.4.0.bom.json
@@ -1,35 +1,35 @@
 {
   "bomFormat": "CycloneDX",
   "specVersion": "1.3",
-  "serialNumber": "urn:uuid:487c5862-7a2c-461a-82f4-9c96e2b3713b",
+  "serialNumber": "urn:uuid:15e86e7c-bd36-40c2-8181-5c1260d23cd1",
   "version": 1,
   "metadata": {
-    "timestamp": "2021-08-06T21:56:56+02:00",
+    "timestamp": "2021-08-08T19:09:43+02:00",
     "tools": [
       {
         "vendor": "CycloneDX",
         "name": "cyclonedx-gomod",
-        "version": "v0.0.0-20210806212106-4c0fcdc6dd84",
+        "version": "v0.0.0-20210806234054-8e6e5d7e4328",
         "hashes": [
           {
             "alg": "MD5",
-            "content": "ce1a46a77c93e8c73327648c34ea3558"
+            "content": "6f45202f726c6e76f1a09b201324056d"
           },
           {
             "alg": "SHA-1",
-            "content": "e8619018a0e7199335e9ad8e7f4336f8dd52f872"
+            "content": "e62130d5e266c0eb9bf65c884e5070ee4d60090b"
           },
           {
             "alg": "SHA-256",
-            "content": "86af05b597621c3b569c6b584c3868cc42d9c790c1378ec832d8bbc04a99b7c7"
+            "content": "a4f31c0d02b0c4a9c900975c9e61f7efaf094c1b9f85656feb0e746cfa187054"
           },
           {
             "alg": "SHA-384",
-            "content": "d2f7aa96a9a5447f4cb6ee58aeba854d8c22b5a044d15e47728139a09afb0b6402867f85888d2aa534c87a4ea5eefb6e"
+            "content": "d42f0f4014a1608467e881c95fa4d3a640c838baf30cead07ef4733784590efb9ea5f20e05d76f2ff4d64f10ea9175fb"
           },
           {
             "alg": "SHA-512",
-            "content": "d49be70a2b99b59c9a2beeda7c09f93c68bf63969234a0e50ab8d3ecc018c87fa73ef84dfe55b41e18b64fc16818a3cd459140d6aca0cf135df414f51d7750f4"
+            "content": "d517e7ffc5a38094431f5a991227a3256600825d5ff46718380b9e6b043a6fb387276a6fec1cf35872ab9b81838bc9993fa3b812cb07a7885c52dec540d5e70d"
           }
         ]
       }
@@ -206,6 +206,13 @@
   ],
   "dependencies": [
     {
+      "ref": "pkg:golang/github.com/CycloneDX/cyclonedx-go@v0.4.0",
+      "dependsOn": [
+        "pkg:golang/github.com/bradleyjkemp/cupaloy/v2@v2.6.0",
+        "pkg:golang/github.com/stretchr/testify@v1.7.0"
+      ]
+    },
+    {
       "ref": "pkg:golang/github.com/bradleyjkemp/cupaloy/v2@v2.6.0",
       "dependsOn": [
         "pkg:golang/github.com/davecgh/go-spew@v1.1.1",
@@ -229,13 +236,6 @@
     },
     {
       "ref": "pkg:golang/gopkg.in/yaml.v3@v3.0.0-20200313102051-9f266ea9e77c"
-    },
-    {
-      "ref": "pkg:golang/github.com/CycloneDX/cyclonedx-go@v0.4.0",
-      "dependsOn": [
-        "pkg:golang/github.com/bradleyjkemp/cupaloy/v2@v2.6.0",
-        "pkg:golang/github.com/stretchr/testify@v1.7.0"
-      ]
     }
   ]
 }
