+++ internal/cli/cmd/mod/testdata/integration/snapshots/TestIntegrationVendored
@@ -28,9 +28,9 @@
     </component>
   </components>
   <dependencies>
-    <dependency ref="pkg:golang/github.com/google/uuid@v1.2.0"></dependency>
     <dependency ref="pkg:golang/testmod-vendored@v0.0.0-20210716205931-5c9f3d791930">
       <dependency ref="pkg:golang/github.com/google/uuid@v1.2.0"></dependency>
     </dependency>
+    <dependency ref="pkg:golang/github.com/google/uuid@v1.2.0"></dependency>
   </dependencies>
 </bom>
